+++ dwarves_fprintf.c
@@ -725,7 +725,7 @@ static size_t struct_member__fprintf(struct class_member *member,
 
 	if (member->is_static) {
 		if (member->const_value != 0)
-			printed += fprintf(fp, " = %lld;", member->const_value);
+			printed += fprintf(fp, " = %lld;", (long long)member->const_value);
 	} else if (member->bitfield_size != 0) {
 		printed += fprintf(fp, ":%u;", member->bitfield_size);
 	} else {
@@ -1459,9 +1459,9 @@ size_t class__fprintf(struct class *class, const struct cu *cu,
 	if (!cconf.emit_stats)
 		goto out;
 
-	printed += fprintf(fp, "\n%.*s/* size: %zd, cachelines: %zd, members: %u",
+	printed += fprintf(fp, "\n%.*s/* size: %zu, cachelines: %zd, members: %u",
 			   cconf.indent, tabs,
-			   class__size(class),
+			   (size_t)class__size(class),
 			   tag__nr_cachelines(class__tag(class), cu),
 			   type->nr_members,
 			   type->nr_static_members);
